Movie rating average

+++ MovieDatabase/Models/Movie.cs
@@ -11,7 +11,7 @@ namespace MovieDatabase.Models
         public string posterImagePath { get; set; }
         public string? description { get; set; }
         public string? trailer_link {  get; set; }
-        public int? rate { get; set; }
+        public float? rate { get; set; }
 
         public ICollection<Actor> actors { get; set;  } = [];
         public ICollection<UserMovie> usermovies { get; set; } = [];
